requirements.txt
allure-pytest==2.8.18
allure_python_commons==2.8.18
elasticsearch==7.13.4
jsonpath==0.82
jsonpath_rw==1.4.0
jsonpath_rw_ext==1.2.2
phoenixdb==1.2.1
pysolr==3.8.1
pytest==7.1.3
PyYAML==6.0
Requests==2.31.0
requests_toolbelt==0.9.1
ruamel.base==1.0.0
simplejson==3.17.2
pytest-cache==1.0
pytest-cov==4.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-forked==1.6.0
pytest-html==3.0.0
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==9.1.1
pytest-xdist==2.1.0